pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "player_control"
version = "0.1.0"
dependencies = [
    "dbus-python",
]
requires-python = ">=3.8"

[project.scripts]
player_control = "player_control.__main__:main"

[project.optional-dependencies]
dev = [
    "pytest",
    "ruff",
]

[tool.ruff]
line-length = 88
target-version = "py38"
